# [perf/UX] Reduce import time