--- a/src/pharmpy/cli.py
+++ b/src/pharmpy/cli.py
@@ -53,54 +53,14 @@ Definitions
 """
 
 import argparse
-import importlib
 import pathlib
 import pydoc
 import sys
-import types
 import warnings
 from collections import OrderedDict, namedtuple
 from textwrap import dedent
 
-
-class LazyLoader(types.ModuleType):
-    """Class that masquerades as a module and lazily loads it when accessed the first time
-
-    The code for the class is taken from TensorFlow and is under the Apache 2.0 license
-
-    This is needed for the cli to be able to do things as version and help quicker than
-    if all modules were imported.
-    """
-
-    def __init__(self, local_name, parent_module_globals, name):
-        self._local_name = local_name
-        self._parent_module_globals = parent_module_globals
-
-        super(LazyLoader, self).__init__(name)
-
-    def _load(self):
-        # Import the target module and insert it into the parent's namespace
-        module = importlib.import_module(self.__name__)
-        self._parent_module_globals[self._local_name] = module
-
-        # Update this object's dict so that if someone keeps a reference to the
-        #   LazyLoader, lookups are efficient (__getattr__ is only called on lookups
-        #   that fail).
-        self.__dict__.update(module.__dict__)
-        return module
-
-    def __getattr__(self, item):
-        module = self._load()
-        return getattr(module, item)
-
-    def __dir__(self):
-        module = self._load()
-        return dir(module)
-
-
-pharmpy = LazyLoader('pharmpy', globals(), 'pharmpy')
-plugin_utils = LazyLoader('plugin_utils', globals(), 'pharmpy.plugins.utils')
-pd = LazyLoader('pd', globals(), 'pandas')
+from .deps import pandas as pd
 
 formatter = argparse.ArgumentDefaultsHelpFormatter
 
@@ -294,9 +254,11 @@ def data_print(args):
 
 def data_filter(args):
     """Subcommand to filter a dataset"""
+    from pharmpy.plugins.utils import PluginError
+
     try:
         df = args.model_or_dataset.dataset
-    except plugin_utils.PluginError:
+    except PluginError:
         df = args.model_or_dataset
     expression = ' '.join(args.expressions)
     try:
@@ -310,9 +272,11 @@ def data_filter(args):
 
 def data_append(args):
     """Subcommand to append a column to a dataset"""
+    from pharmpy.plugins.utils import PluginError
+
     try:
         df = args.model_or_dataset.dataset
-    except plugin_utils.PluginError:
+    except PluginError:
         df = args.model_or_dataset
     try:
         df = df.eval(args.expression)
@@ -406,6 +370,8 @@ def data_anonymize(args):
 
 def info(args):
     """Subcommand to print Pharmpy info (and brag a little bit)."""
+
+    import pharmpy
 
     Install = namedtuple('VersionInfo', ['version', 'authors', 'directory'])
     inst = Install(
@@ -866,15 +832,21 @@ def input_model(path):
     Raises if not found or is dir, without tracebacks (see :func:`error_exit`).
     """
     path = check_input_path(path)
-    return pharmpy.Model.create_model(path)
+    from pharmpy.model import Model
+
+    return Model.create_model(path)
 
 
 def input_model_or_dataset(path):
     """Returns :class:`~pharmpy.model.Model` or pd.DataFrame from *path*"""
     path = check_input_path(path)
+    from pharmpy.plugins.utils import PluginError
+
     try:
-        obj = pharmpy.Model.create_model(path)
-    except pharmpy.plugins.utils.PluginError:
+        from pharmpy.model import Model
+
+        obj = Model.create_model(path)
+    except PluginError:
         obj = pd.read_csv(path)
     return obj