--- a/src/pharmpy/deps/__init__.py
+++ b/src/pharmpy/deps/__init__.py
@@ -1,9 +1,17 @@
-from .lazyload import LazyLoader
+from ..lazyload import LazyLoader
 
-altair = LazyLoader('altair', globals(), 'altair')
+altair = LazyLoader('altair', globals(), 'pharmpy.deps.altair', 'altair')
 networkx = LazyLoader('networkx', globals(), 'networkx')
 numpy = LazyLoader('numpy', globals(), 'numpy')
 pandas = LazyLoader('pandas', globals(), 'pandas')
-scipy = LazyLoader('scipy', globals(), 'scipy')
 symengine = LazyLoader('symengine', globals(), 'symengine')
 sympy = LazyLoader('sympy', globals(), 'sympy')
+
+__all__ = (
+    'altair',
+    'networkx',
+    'numpy',
+    'pandas',
+    'symengine',
+    'sympy',
+)