--- a/src/pharmpy/deps/rich.py
+++ b/src/pharmpy/deps/rich.py
@@ -1,0 +1,5 @@
+from ..lazyload import LazyLoader
+
+box = LazyLoader('box', globals(), 'rich.box')
+console = LazyLoader('console', globals(), 'rich.console')
+table = LazyLoader('table', globals(), 'rich.table')