--- a/src/pharmpy/deps/scipy.py
+++ b/src/pharmpy/deps/scipy.py
@@ -1,0 +1,4 @@
+from ..lazyload import LazyLoader
+
+stats = LazyLoader('stats', globals(), 'scipy.stats')
+linalg = LazyLoader('linalg', globals(), 'scipy.linalg')