--- a/src/pharmpy/expressions.py
+++ b/src/pharmpy/expressions.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from functools import reduce
 from itertools import chain
 from operator import __and__