--- a/src/pharmpy/modeling/common.py
+++ b/src/pharmpy/modeling/common.py
@@ -7,10 +7,10 @@ import warnings
 from io import StringIO
 from pathlib import Path
 
-import sympy
-
-from pharmpy import Model, Parameter, Parameters, RandomVariables, config
+import pharmpy.config as config
+from pharmpy.deps import sympy
 from pharmpy.modeling import split_joint_distribution
+from pharmpy.objects import Model, Parameter, Parameters, RandomVariables
 from pharmpy.statements import Assignment, CompartmentalSystem
 from pharmpy.utils import normalize_user_given_path
 from pharmpy.workflows import default_model_database