> @rikardn REVIEW

Fixes #1006. This also defines an action that can be triggered manually to check for import times regressions.

Ideas and questions:
  - [x] Maybe use [`from __future__ import annotations`](https://docs.python.org/3/library/__future__.html) instead of removing annotations? ~No since dynamic type annotations are useless anyway.~ Yes, because the code looks better and will be ready to migrate to better implementation in Python 3.11.
  - [x] Move stuff from `pharmpy.objects` to `pharmpy.model`. ~More thought needs to be put into where the separation between `pharmpy.model` and `pharmpy.modeling` would lie.~
  - [x] When should the new `importtime` GHA workflow be run? Running it on all pushes to `main` seems excessive. We could also limit ourselves to a single Python version. **EDIT**: Only manual execution for now.

On my machine:
  - `pharmpy --version` takes less than 105 ms.
  - `from pharmpy.modeling import *` takes less than 80 ms.
  - `from pharmpy.tools import *` takes less than 135 ms.
  - `from pharmpy.tools import run_amd` takes less than 85 ms.
  - `from pharmpy.model import *` takes less than 60 ms.
  - `import pharmpy` takes less than 35 ms.
  - `from pharmpy import cli` takes less than 80 ms.
  - `print("Hello, world")` records between 25ms and 35ms import time.

## Some flame graphs
  
### `from pharmpy.modeling import *`

![Screenshot 2022-09-05 at 16 35 23](https://user-images.githubusercontent.com/101794402/188473796-4dc9fd17-e073-4e62-a37a-bdeb9e435d43.png)

### `from pharmpy.tools import *`

![Screenshot 2022-09-05 at 16 35 38](https://user-images.githubusercontent.com/101794402/188473878-789960f0-bca0-49a3-ab79-aa134f7a82c7.png)

### `from pharmpy.tools import run_amd`

![Screenshot 2022-09-05 at 16 36 39](https://user-images.githubusercontent.com/101794402/188474000-b9debb47-161f-4174-81a4-e7c39220c282.png)

### `from pharmpy.model import *`

![Screenshot 2022-09-05 at 16 35 00](https://user-images.githubusercontent.com/101794402/188473733-57854d21-a9d8-4368-8ce4-9275502a5963.png)

### `import pharmpy`

![Screenshot 2022-09-05 at 16 35 56](https://user-images.githubusercontent.com/101794402/188473947-43e8ac70-a279-4496-9a97-a49f0ee3252a.png)

### `from pharmpy import cli`

![Screenshot 2022-09-05 at 16 34 17](https://user-images.githubusercontent.com/101794402/188473677-679d29b9-6555-4f8d-823f-1961df198597.png)